Fix a view file button to can not use
