+++ kuhub/static/kuhub/js/config.js
@@ -1,0 +1,28 @@
+const firebaseConfig = {
+    apiKey: process.env.FIREBASE_API_KEY,
+    authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+    projectId: process.env.FIREBASE_PROJECT_ID,
+    storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+    messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
+    appId: process.env.FIREBASE_APP_ID,
+    measurementId: process.env.FIREBASE_MEASUREMENT_ID
+};
+
+export default firebaseConfig;
+
+// app.get('/', (req, res) => {
+//     res.render('index', {
+//         firebaseConfig: JSON.stringify({
+//             apiKey: process.env.FIREBASE_API_KEY,
+//             authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+//             projectId: process.env.FIREBASE_PROJECT_ID,
+//             storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+//             messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
+//             appId: process.env.FIREBASE_APP_ID,
+//             measurementId: process.env.FIREBASE_MEASUREMENT_ID
+//         })
+//     });
+// });
+
+
+// export default firebaseConfig;
