+++ kuhub/templates/kuhub/summary.html
@@ -120,7 +120,7 @@
                     <br>
                     <p class="card-text">{{ post.post_id.post_content }}</p>
                     <p class="card-text">
-                        <button onclick="viewFile('{{ post.file.name }}')">View File</button>
+                        <button>View File</button>
                     </p>
                     <h6 class="card-subtitle mb-2 text-muted"><i class="fa-regular fa-clock" style="margin-right: 5px;"></i>{{ post.post_id.post_date }}</h6>
